Change NutritionFact.name to be a read only property.
Problem can occur when working with NutritionFacts and
NutritionSets with the name attribute of NutritionFact.
NutritionSets are dictionaries that are keyed based on the
NutritionFact's name. When the NutritionFact name attribute
is not read-only, it becomes possible for the key of the
NutritionSet to be different than than the name of the NutritionFact
that the key maps to.