Add support for adding and subtracting NutritionSet instances.
Did a lot of refactoring with the NutritionSet.update method in order
to get this to work. Added the parameter merge_func to the update
method which allows one to pass a function that will be used
to merge two of the same NutritionFacts if they are both present
in each NutritionSet. Additionally fixed bug where if a NutritionSet
doesn't have a NutritionFact, it will return a new NutritionFact with
a dimensionless quantity with magnitue zero, which can't be added
to other NutritionFacts with non-dimensionless quantities. Solution
was found by modifying NutritionFact.__(add/sub)__ to explicitly
solve for dimensionless quantities with magnitute zero.